pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openapi-mcp-generator"
version = "0.1.0"
description = "Generate MCP servers from OpenAPI specifications"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1.2",
    "httpx>=0.24.0",
]

[project.scripts]
mcp-generator = "openapi_mcp_generator.cli:main"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 100
